+++ ikuzo/service/x/index/option.go
@@ -16,7 +16,9 @@ package index
 
 import (
 	"github.com/delving/hub3/ikuzo/domain"
+	"github.com/delving/hub3/ikuzo/service/organization"
 	"github.com/elastic/go-elasticsearch/v8/esutil"
+	"github.com/rs/zerolog"
 )
 
 type Option func(*Service) error
@@ -53,12 +55,33 @@ func SetOrphanWait(wait int) Option {
 	}
 }
 
+func SetDisableMetrics(disable bool) Option {
+	return func(s *Service) error {
+		s.disableMetrics = disable
+		return nil
+	}
+}
+
+func SetOrganisationService(org *organization.Service) Option {
+	return func(s *Service) error {
+		s.orgs = org
+		return nil
+	}
+}
+
 func SetPostHookService(hooks ...domain.PostHookService) Option {
 	return func(s *Service) error {
 		for _, hook := range hooks {
 			s.postHooks[hook.OrgID()] = append(s.postHooks[hook.OrgID()], hook)
 		}
 
+		return nil
+	}
+}
+
+func SetLogger(log zerolog.Logger) Option {
+	return func(s *Service) error {
+		s.log = log
 		return nil
 	}
 }